Use cobra for argument parsing.
