+++ go.mod
@@ -6,6 +6,7 @@ require (
 	cloud.google.com/go/container v1.43.0
 	cloud.google.com/go/recommender v1.13.5
 	github.com/mark3labs/mcp-go v0.32.0
+	github.com/spf13/cobra v1.9.1
 	google.golang.org/api v0.233.0
 	google.golang.org/protobuf v1.36.6
 )
@@ -21,7 +22,9 @@ require (
 	github.com/google/uuid v1.6.0 // indirect
 	github.com/googleapis/enterprise-certificate-proxy v0.3.6 // indirect
 	github.com/googleapis/gax-go/v2 v2.14.2 // indirect
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
 	github.com/spf13/cast v1.7.1 // indirect
+	github.com/spf13/pflag v1.0.6 // indirect
 	github.com/yosida95/uritemplate/v3 v3.0.2 // indirect
 	go.opentelemetry.io/auto/sdk v1.1.0 // indirect
 	go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc/otelgrpc v0.60.0 // indirect
